*** empty log message ***
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@5093 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/promemul/dec_prom.cc
@@ -25,7 +25,7 @@
  *  SUCH DAMAGE.
  *
  *
- *  $Id: dec_prom.cc,v 1.1 2007-11-17 12:13:53 debug Exp $
+ *  $Id: dec_prom.cc,v 1.2 2007-11-17 12:42:23 debug Exp $
  *
  *  COMMENT: DECstation PROM emulation
  *
@@ -132,7 +132,7 @@ int dec_jumptable_func(struct cpu *cpu, int vector)
 		cpu->cd.mips.gpr[MIPS_GPR_V0] = 1;
 		break;
 	case 0x38:	/*  read(handle, ptr, length)  */
-		cpu->cd.mips.gpr[MIPS_GPR_V0] = -1;
+		cpu->cd.mips.gpr[MIPS_GPR_V0] = (int64_t) -1;
 		if ((int32_t)cpu->cd.mips.gpr[MIPS_GPR_A2] > 0) {
 			int disk_id = diskimage_bootdev(cpu->machine, NULL);
 			int res;
@@ -260,7 +260,7 @@ int decstation_prom_emul(struct cpu *cpu)
 		if ((signed char)ch1 > (signed char)ch2)
 			cpu->cd.mips.gpr[MIPS_GPR_V0] = 1;
 		if ((signed char)ch1 < (signed char)ch2)
-			cpu->cd.mips.gpr[MIPS_GPR_V0] = -1;
+			cpu->cd.mips.gpr[MIPS_GPR_V0] = (int64_t) -1;
 		break;
 	case 0x14:		/*  strlen():  */
 		i = 0;
